sending datagrams from client

+++ clock.cpp
@@ -1,7 +1,7 @@
 #include "clock.h"
 #include "utils.h"
 
-Clock::Clock(uint32_t turn_time): next_turn_time(turn_time), average_turn_time(turn_time) {
+Clock::Clock(uint64_t turn_time): next_turn_time(turn_time), average_turn_time(turn_time) {
     last_turn_start_time = get_time_microseconds();
 }
 
@@ -13,7 +13,7 @@ bool Clock::end_turn() const {
 
 void Clock::next_turn() {
     uint64_t temp = get_time_microseconds();
-    uint32_t passed = (uint32_t)(temp - last_turn_start_time); /* used time */
-    int32_t delta = next_turn_time - passed; /* how much left from the previous turn, may be negative */
+    uint64_t passed = temp - last_turn_start_time; /* used time */
+    int64_t delta = next_turn_time - passed; /* how much left from the previous turn, may be negative */
     next_turn_time = average_turn_time + delta;
 }
